+++ bigdata/src/java/com/bigdata/service/EmbeddedFederation.java
@@ -41,21 +41,44 @@ import com.bigdata.journal.BufferMode;
 import com.bigdata.journal.IResourceLockService;
 import com.bigdata.journal.ITransactionService;
 import com.bigdata.journal.ResourceLockService;
+import com.bigdata.journal.WriteExecutorService;
 import com.bigdata.service.EmbeddedClient.Options;
 
 /**
  * An implementation that uses an embedded database rather than a distributed
  * database. An embedded federation runs entirely in process, but uses the same
  * {@link DataService} and {@link MetadataService} implementations as a
- * distributed federation. Unlike a distributed federation, an embedded
- * federation starts and stops with the client. An embedded federation may be
- * used to assess or remove the overhead of network operations, to simplify
- * testing of client code, or to deploy a scale-up (vs scale-out) solution.
+ * distributed federation. All services reference the {@link EmbeddedFederation}
+ * and use the same thread pool for most operations. However, the
+ * {@link EmbeddedDataServiceImpl} has its own {@link WriteExecutorService}.
+ * Unlike a distributed federation, an embedded federation starts and stops with
+ * the client. An embedded federation may be used to assess or remove the
+ * overhead of network operations, to simplify testing of client code, or to
+ * deploy a scale-up (vs scale-out) solution.
  * 
  * @author <a href="mailto:thompsonbry@users.sourceforge.net">Bryan Thompson</a>
  * @version $Id$
+ * 
+ @todo The EDS/LDS should use their own options in their own namespace to
+ *       specify the data directory for the federation. Ditto for the
+ *       "transient" or "createTempFile" properties. Everything is namespaced
+ *       now and the overridden semantics of
+ *       com.bigdata.journal.Options.CREATE_TEMP_FILE and StoreManager#DATA_DIR
+ *       are just getting us into trouble. Look at all uses of these options in
+ *       the unit tests and decouple them from the journal's options.
  */
 public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
+
+    /**
+     * Text of the warning message used when a file or directory could not be
+     * deleted during {@link #destroy()}.
+     */
+    private static final String ERR_COULD_NOT_DELETE = "Could not delete: ";
+
+    /**
+     * The name of the file used to mark an MDS vs DS service.
+     */
+    static private final String MDS = ".mds";
 
     /**
      * The #of data service instances.
@@ -64,7 +87,6 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
     
     /**
      * True if the federation is not backed by disk.
-     * @return
      */
     private final boolean isTransient;
     
@@ -75,25 +97,39 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
     private final File dataDir;
     
     /**
+     * The directory in which the data files will reside. Each directory
+     * is named for the service {@link UUID} - restart depends on this.
+     */
+    public final File getDataDir() {
+        
+        return dataDir;
+        
+    }
+    
+    /**
      * The (in process) {@link AbstractTransactionService}.
      */
-    private AbstractTransactionService abstractTransactionService;
+    private final AbstractTransactionService abstractTransactionService;
     
     /** The (in process) {@link IResourceLockService} */
-    private ResourceLockService resourceLockManager;
+    private final ResourceLockService resourceLockManager;
     
     /**
      * The (in process) {@link LoadBalancerService}.
      */
-    private LoadBalancerService loadBalancerService;
+    private final LoadBalancerService loadBalancerService;
     
     /**
      * The (in process) {@link MetadataService}.
+     * <p>
+     * Note: Not final because not initialized in the constructor.
      */
     private MetadataService metadataService;
     
     /**
      * The (in process) {@link DataService}s.
+     * <p>
+     * Note: Not final because not initialized in the constructor.
      */
     private DataService[] dataService;
     
@@ -168,7 +204,7 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
      * @return The {@link DataService} for that UUID or <code>null</code> if
      *         there is no data service instance with that service UUID.
      */
-    final public IDataService getDataService(UUID serviceUUID) {
+    final public IDataService getDataService(final UUID serviceUUID) {
 
         // Note: return null if service not available/discovered.
 
@@ -372,11 +408,39 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
          */
         resourceLockManager = new ResourceLockService();
         
-        /*
-         * Start the load balancer.
-         */
-        loadBalancerService = new EmbeddedLoadBalancerServiceImpl(UUID.randomUUID(),
-                properties).start();
+        {
+
+            final Properties p = new Properties(properties);
+            
+            if (isTransient) {
+
+                p.setProperty(LoadBalancerService.Options.TRANSIENT, "true");
+
+            } else {
+                
+                // specify the data directory for the load balancer.
+                p.setProperty(EmbeddedLoadBalancerServiceImpl.Options.LOG_DIR,
+                        new File(dataDir, "lbs").toString());
+                
+            }
+
+            /*
+             * Start the load balancer.
+             */
+            try {
+         
+                loadBalancerService = new EmbeddedLoadBalancerServiceImpl(UUID
+                        .randomUUID(), p).start();
+            
+            } catch (Throwable t) {
+            
+                log.error(t, t);
+                
+                throw new RuntimeException(t);
+                
+            }
+
+        }
 
         /*
          * The directory in which the data files will reside.
@@ -470,9 +534,9 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
                     p.setProperty(MetadataService.Options.DATA_DIR, serviceDir.toString());
 //                    p.setProperty(Options.FILE, new File(serviceDir,"journal"+Options.JNL).toString());
 
-                    if(new File(serviceDir,".mds").exists()) {
+                    if(new File(serviceDir,MDS).exists()) {
                         
-                        /*
+                        /*`
                          * metadata service.
                          */
                         metadataService = new EmbeddedMetadataService(this,
@@ -571,10 +635,11 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
      * are requested. We wind up creating the resource manager files in the
      * current working directory when we really want the resource manager to
      * startup with transient journals (and disallow overflow since you can not
-     * re-open a store or even write an index segement).
+     * re-open a store or even write an index segment).
      */
-    private int createFederation(Properties properties, boolean isTransient) {
-        
+    private int createFederation(final Properties properties,
+            final boolean isTransient) {
+
         final int ndataServices;
         
         /*
@@ -582,7 +647,7 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
          */
         {
 
-            String val = properties.getProperty(Options.NDATA_SERVICES,
+            final String val = properties.getProperty(Options.NDATA_SERVICES,
                     Options.DEFAULT_NDATA_SERVICES);
 
             ndataServices = Integer.parseInt(val);
@@ -617,7 +682,7 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
                  */
                 try {
 
-                    new RandomAccessFile(new File(serviceDir, ".mds"), "rw")
+                    new RandomAccessFile(new File(serviceDir, MDS), "rw")
                             .close();
 
                 } catch (IOException e) {
@@ -692,7 +757,7 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
         }
 
         @Override
-        public EmbeddedFederation getFederation() {
+        public EmbeddedFederation<T> getFederation() {
 
             return EmbeddedFederation.this;
 
@@ -713,7 +778,7 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
         }
 
         @Override
-        public EmbeddedFederation getFederation() {
+        public EmbeddedFederation<T> getFederation() {
 
             return EmbeddedFederation.this;
 
@@ -734,7 +799,7 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
         }
 
         @Override
-        public EmbeddedFederation getFederation() {
+        public EmbeddedFederation<T> getFederation() {
 
             return EmbeddedFederation.this;
 
@@ -771,24 +836,30 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
 
         for (int i = 0; i < dataService.length; i++) {
 
-            DataService ds = this.dataService[i];
+            if (dataService[i] != null) {
 
-            ds.shutdownNow();
+                dataService[i].shutdown();
+
+            }
             
         }
 
-        metadataService.shutdownNow();
+        if (metadataService != null) {
+
+            metadataService.shutdown();
+            
+        }
         
         if (loadBalancerService != null) {
 
             loadBalancerService.shutdown();
 
-            loadBalancerService = null;
+//            loadBalancerService = null;
             
         }
         
-        // Note: don't clear ref until all down since nextTimestamp() still active.
-        abstractTransactionService = null;
+//        // Note: don't clear ref until all down since nextTimestamp() still active.
+//        abstractTransactionService = null;
 
         if (log.isInfoEnabled())
             log.info("done");
@@ -813,10 +884,12 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
 
         for (int i = 0; i < dataService.length; i++) {
 
-            DataService ds = this.dataService[i];
-            
-            ds.shutdownNow();
-            
+            if (dataService[i] != null) {
+
+                dataService[i].shutdownNow();
+
+            }
+
         }
 
         metadataService.shutdownNow();
@@ -825,12 +898,12 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
 
             loadBalancerService.shutdownNow();
 
-            loadBalancerService = null;
+//            loadBalancerService = null;
             
         }
         
-        // Note: don't clear ref until all down since nextTimestamp() still active.
-        abstractTransactionService = null;
+//        // Note: don't clear ref until all down since nextTimestamp() still active.
+//        abstractTransactionService = null;
 
         if (log.isInfoEnabled())
             log.info("done");
@@ -839,51 +912,46 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
 
     public void destroy() {
 
-        if (log.isInfoEnabled())
-            log.info("");
+        super.destroy();
 
-        abstractTransactionService.shutdownNow();
+        abstractTransactionService.destroy();
         
         for (int i = 0; i < dataService.length; i++) {
 
-            IDataService ds = dataService[i];
-            
-            try {
-                
-                ds.destroy();
-            
-            } catch (IOException e) {
-             
-                log.error("Could not destroy dataService", e );
+            if (dataService[i] != null) {
+
+                dataService[i].destroy();
                 
             }
-            
-            dataService[i] = null;
-
+ 
         }
 
-        {
+        if (metadataService != null) {
 
-            try {
+            // the file flagging this as the MDS rather than a DS.
+            final File tmp = new File(metadataService.getResourceManager()
+                    .getDataDir(), EmbeddedFederation.MDS);
 
-                metadataService.destroy();
+            if(!tmp.delete()) {
 
-            } catch (IOException e) {
-
-                log.error("Could not destroy dataService", e);
+                log.warn(ERR_COULD_NOT_DELETE + tmp);
 
             }
 
-            metadataService = null;
-            
+            metadataService.destroy();
+
         }
 
-        loadBalancerService.shutdownNow();
+        loadBalancerService.destroy();
+
+        if (!isTransient && !dataDir.delete()) {
+
+            log.warn(ERR_COULD_NOT_DELETE + dataDir);
+            
+        }
         
-        loadBalancerService = null;
-
-        // Note: don't clear ref until all down since nextTimestamp() still active.
-        abstractTransactionService = null;
+//        // Note: don't clear ref until all down since nextTimestamp() still active.
+//        abstractTransactionService = null;
         
     }
